2003-12-29  Bob Doan  <bdoan@sicompos.com> 	* bindings/php/php.c: use glib data types 	* inputs/mysql/mysql.c: likewise 	* inputs/odbc/odbc.c: likewise 	* inputs/postgre/postgre.c: likewise 	* libsrc/api.c: likewise 	* libsrc/breaks.c: likewise 	* libsrc/csv.c: likewise 	* libsrc/datasource.c: likewise 	* libsrc/environment.c: likewise 	* libsrc/formatstring.c: likewise 	* libsrc/free.c: likewise 	* libsrc/fxp.c: likewise 	* libsrc/html.c: likewise 	* libsrc/navigation.c: likewise 	* libsrc/parsexml.c: likewise 	* libsrc/pcode.c: likewise 	* libsrc/pcode_op_functions.c: likewise 	* libsrc/pdf.c: likewise 	* libsrc/ralloc.c: likewise 	* libsrc/reportgen.c: likewise 	* libsrc/resolution.c: likewise 	* libsrc/rhashtable.c: likewise 	* libsrc/rhashtable.h:likewise 	* libsrc/rlib.h: likewise 	* libsrc/rlib_input.h: likewise 	* libsrc/rvector.c: likewise 	* libsrc/rvector.h: likewise 	* libsrc/txt.c: likewise 	* libsrc/util.c: likewise

+++ libsrc/fxp.c
@@ -20,20 +20,20 @@
  
 #include <stdio.h>
 #include <stdlib.h>
-
+#include <glib.h>
 
 #define DIV0(a,b) ((b) == 0 ? 0 : ((a)/(b)))
 #define MOD0(a,b) ((b) == 0 ? 0 : ((a)%(b)))
 
 //man 3 llabs says the prototype is in stdlib.. no it aint!
-long long int llabs(long long int j);
+gint64 llabs(gint64 j);
 
 
-long long rlib_fxp_mul(long long a, long long b, long long factor) {
-	long long wholea;
-	long long wholeb;
-	long long parta;
-	long long partb;
+gint64 rlib_fxp_mul(gint64 a, gint64 b, gint64 factor) {
+	gint64 wholea;
+	gint64 wholeb;
+	gint64 parta;
+	gint64 partb;
 
 	wholea = DIV0(a,factor);
 	wholeb = DIV0(b,factor);
@@ -42,8 +42,8 @@ long long rlib_fxp_mul(long long a, long long b, long long factor) {
 	return( (wholea*wholeb)*factor+(parta*wholeb+partb*wholea)+DIV0(parta*partb,factor));
 }
 
-long long rlib_fxp_div( long long num, long long denom, int places) {
-	long long result = 0l;
+gint64 rlib_fxp_div(gint64 num, gint64 denom, gint places) {
+	gint64 result = 0l;
 
 	if( denom != 0L) {
 		if( (llabs(denom)&0xFC00000000000000LL) ) {
